+++ src/eg3dmath.c
@@ -19,7 +19,7 @@
 /* sin/cos via table lookup + linear interpolation.
  * Mirrors sineLookup: idx = angle>>8, frac = angle&0xFF, then
  *   result = lut[idx] + round((lut[idx+1] - lut[idx]) * frac / 256). */
-static int16 sineInterp(unsigned angle) {
+static int16 sineInterp(uint16 angle) {
     int16 idx = (angle >> 8) & 0xFF;
     int16 frac = angle & 0xFF;
     int16 v0 = g_angleLut[idx];
@@ -34,7 +34,7 @@ int16 sine(int16 angle) {
 
 int16 cosine(int16 angle) {
     /* cos(x) = sin(x + 90deg); a quarter turn is 0x4000 in 16-bit angle space. */
-    return sineInterp((unsigned)angle + 0x4000);
+    return sineInterp((uint16)angle + 0x4000);
 }
 
 /* Q15-style fixed multiply: returns round((a*b) >> 15).
